chore: 구간 인수 테스트를 별도의 클래스로 분리
