+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -6,7 +6,6 @@ import static nextstep.subway.station.StationFixture.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -23,7 +22,6 @@ import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineCreateRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.LineUpdateRequest;
-import nextstep.subway.line.dto.SectionAddRequest;
 import nextstep.subway.station.dto.StationResponse;
 
 @DisplayName("지하철 노선 관련 기능")
@@ -177,175 +175,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         지하철_노선_찾지_못함(response);
     }
 
-    /**
-     * [AS-IS] (강남역)---7m----(선릉역)
-     * [ ADD ] (강남역)--4m--(역삼역)
-     * [TO-BE] (강남역)--4m--(역삼역)-3m--(선릉역)
-     */
-    @DisplayName("지하철 구간을 등록한다. (역 사이에 새로운 역을 등록할 경우)")
-    @Test
-    void addSectionBetweenStations() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 선릉역_생성_응답.getId(), 7);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 4);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록됨(지하철_구간_등록_응답);
-        지하철_노선에_지하철_역이_순서대로_등록됨(노선_2호선_생성_응답.getId(), Arrays.asList(
-            강남역_생성_응답.getId(),
-            역삼역_생성_응답.getId(),
-            선릉역_생성_응답.getId()));
-    }
-
-    /**
-     * [AS-IS]             (역삼역)---7m----(선릉역)
-     * [ ADD ] (강남역)--4m--(역삼역)
-     * [TO-BE] (강남역)--4m--(역삼역)---7m----(선릉역)
-     */
-    @DisplayName("지하철 구간을 등록한다. (새로운 역을 상행 종점으로 등록할 경우)")
-    @Test
-    void addSectionToUpStation() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(역삼역_생성_응답.getId(), 선릉역_생성_응답.getId(), 7);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 4);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록됨(지하철_구간_등록_응답);
-        지하철_노선에_지하철_역이_순서대로_등록됨(노선_2호선_생성_응답.getId(), Arrays.asList(
-            강남역_생성_응답.getId(),
-            역삼역_생성_응답.getId(),
-            선릉역_생성_응답.getId()));
-    }
-
-    /**
-     * [AS-IS] (강남역)---7m----(역삼역)
-     * [ ADD ]                (역삼역)-3m--(선릉역)
-     * [TO-BE] (강남역)---7m----(역삼역)-3m--(선릉역)
-     *
-     */
-    @DisplayName("지하철 구간을 등록한다. (새로운 역을 하행 종점으로 등록할 경우)")
-    @Test
-    void addSectionToDownStation() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 7);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(역삼역_생성_응답.getId(), 선릉역_생성_응답.getId(), 3);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록됨(지하철_구간_등록_응답);
-        지하철_노선에_지하철_역이_순서대로_등록됨(노선_2호선_생성_응답.getId(), Arrays.asList(
-            강남역_생성_응답.getId(),
-            역삼역_생성_응답.getId(),
-            선릉역_생성_응답.getId()));
-    }
-
-    @DisplayName("상행역과 하행역 둘 중 하나도 포함되어 있지 않을 때, 구간을 등록할 수 없다.")
-    @Test
-    void addSectionFailOnBothStationsNotRegistered() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-        StationResponse 삼성역_생성_응답 = 지하철_역_등록되어_있음(삼성역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 1);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(선릉역_생성_응답.getId(), 삼성역_생성_응답.getId(), 2);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록_실패됨(지하철_구간_등록_응답);
-    }
-
-    @DisplayName("상행역과 하행역 둘 중 하나도 포함되어 있지 않을 때, 구간을 등록할 수 없다.")
-    @Test
-    void addSectionFailOnBothStationsAlreadyRegistered() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 1);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 2);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록_실패됨(지하철_구간_등록_응답);
-    }
-
-    @DisplayName("역 사이 왼쪽에 새로운 구간을 등록할 때, 기존 역 사이 길이보다 크거나 같으면 등록 할 수 없다.")
-    @Test
-    void addSectionBetweenStationsAtLeftSideFailOnIllegalDistance() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 선릉역_생성_응답.getId(), 7);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(강남역_생성_응답.getId(), 역삼역_생성_응답.getId(), 7);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록_실패됨(지하철_구간_등록_응답);
-    }
-
-    @DisplayName("역 사이 오른쪽에 새로운 구간을 등록할 때, 기존 역 사이 길이보다 크거나 같으면 등록 할 수 없다.")
-    @Test
-    void addSectionBetweenStationsAtRightSideFailOnIllegalDistance() {
-        // given
-        StationResponse 강남역_생성_응답 = 지하철_역_등록되어_있음(강남역_생성_요청값());
-        StationResponse 역삼역_생성_응답 = 지하철_역_등록되어_있음(역삼역_생성_요청값());
-        StationResponse 선릉역_생성_응답 = 지하철_역_등록되어_있음(선릉역_생성_요청값());
-
-        LineCreateRequest 노선_2호선_생성_요청값 = 노선_2호선_생성_요청값(강남역_생성_응답.getId(), 선릉역_생성_응답.getId(), 7);
-        LineResponse 노선_2호선_생성_응답 = 지하철_노선_등록되어_있음(노선_2호선_생성_요청값);
-
-        SectionAddRequest 구간_등록_요청값 = SectionFixture.구간_등록_요청값(역삼역_생성_응답.getId(), 선릉역_생성_응답.getId(), 7);
-
-        // when
-        ExtractableResponse<Response> 지하철_구간_등록_응답 = 지하철_구간_등록_요청(노선_2호선_생성_응답.getId(), 구간_등록_요청값);
-
-        // then
-        지하철_구간_등록_실패됨(지하철_구간_등록_응답);
-    }
-
-    private ExtractableResponse<Response> 지하철_노선_생성_요청(LineCreateRequest request) {
+    private static ExtractableResponse<Response> 지하철_노선_생성_요청(LineCreateRequest request) {
         return RestAssured.given().log().all()
             .body(request)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -376,7 +206,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         assertThat(actualIds).isEqualTo(expectIds);
     }
 
-    private LineResponse 지하철_노선_등록되어_있음(LineCreateRequest request) {
+    public static LineResponse 지하철_노선_등록되어_있음(LineCreateRequest request) {
         ExtractableResponse<Response> response = 지하철_노선_생성_요청(request);
         return response.jsonPath().getObject(".", LineResponse.class);
     }
@@ -411,7 +241,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         assertThat(actualLineIds).containsAll(expectedLineIds);
     }
 
-    private ExtractableResponse<Response> 지하철_노선_조회_요청(Long id) {
+    public static ExtractableResponse<Response> 지하철_노선_조회_요청(Long id) {
         return RestAssured.given().log().all()
             .when()
             .get("/lines/{lineId}", id)
@@ -452,34 +282,5 @@ public class LineAcceptanceTest extends AcceptanceTest {
 
     private void 지하철_노선_삭제됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-    private ExtractableResponse<Response> 지하철_구간_등록_요청(Long lineId, SectionAddRequest request) {
-        return RestAssured.given().log().all()
-            .body(request)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when()
-            .post("/lines/{lineId}/sections", lineId)
-            .then().log().all()
-            .extract();
-    }
-
-    private void 지하철_구간_등록됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    private void 지하철_노선에_지하철_역이_순서대로_등록됨(Long lineId, List<Long> expectedStationIds) {
-        ExtractableResponse<Response> response = 지하철_노선_조회_요청(lineId);
-        LineResponse lineResponse = response.as(LineResponse.class);
-        List<Long> actualStationIds = lineResponse.getStations()
-            .stream()
-            .map(StationResponse::getId)
-            .collect(Collectors.toList());
-
-        assertThat(actualStationIds).isEqualTo(expectedStationIds);
-    }
-
-    private void 지하철_구간_등록_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 }
